Validaciones login v1.2
Mover funciones de validación a su archivo propio en la carpeta de middlewares
