+++ logLosRoperos.txt
@@ -242,3 +242,19 @@ Visito: /css/styles.css Fri Feb 18 2022 00:22:57 GMT-0600 (hora estándar centra
 Visito: /css/cssLogin.css Fri Feb 18 2022 00:22:57 GMT-0600 (hora estándar central)
 Visito: /images/logo.jpg Fri Feb 18 2022 00:22:57 GMT-0600 (hora estándar central)
 Visito: /images/bike.jpg Fri Feb 18 2022 00:22:57 GMT-0600 (hora estándar central)
+Visito: / Sun Mar 06 2022 17:28:41 GMT-0600 (hora estándar central)
+Visito: /css/home.css Sun Mar 06 2022 17:28:42 GMT-0600 (hora estándar central)
+Visito: /css/bootstrap-grid.min.css Sun Mar 06 2022 17:28:42 GMT-0600 (hora estándar central)
+Visito: /css/styles.css Sun Mar 06 2022 17:28:42 GMT-0600 (hora estándar central)
+Visito: /images/logo.jpg Sun Mar 06 2022 17:28:42 GMT-0600 (hora estándar central)
+Visito: /images/products/1.jpg Sun Mar 06 2022 17:28:42 GMT-0600 (hora estándar central)
+Visito: /favicon.ico Sun Mar 06 2022 17:28:44 GMT-0600 (hora estándar central)
+Visito: /user/login Sun Mar 06 2022 17:28:46 GMT-0600 (hora estándar central)
+Visito: /css/cssLogin.css Sun Mar 06 2022 17:28:46 GMT-0600 (hora estándar central)
+Visito: /css/styles.css Sun Mar 06 2022 17:28:46 GMT-0600 (hora estándar central)
+Visito: /images/logo.jpg Sun Mar 06 2022 17:28:46 GMT-0600 (hora estándar central)
+Visito: /images/bike.jpg Sun Mar 06 2022 17:28:46 GMT-0600 (hora estándar central)
+Visito: /user/login Sun Mar 06 2022 17:28:54 GMT-0600 (hora estándar central)
+Visito: /css/styles.css Sun Mar 06 2022 17:28:55 GMT-0600 (hora estándar central)
+Visito: /css/cssLogin.css Sun Mar 06 2022 17:28:55 GMT-0600 (hora estándar central)
+Visito: /images/logo.jpg Sun Mar 06 2022 17:28:55 GMT-0600 (hora estándar central)
